Example Sound File

A single calling song of *Coryphoda albidicollis*, a Phaneropterinae from Central Chile.

coryphoda

Format

An object of class Wave. The Wave object contains:

left

The left audio channel (if stereo).

samp.rate

The sampling rate in Hz.

bit

The bit depth (e.g., 16 for CD quality).

duration

The total duration of the sound file in seconds.

Source

Original recording by Francisco Rivas Fuenzalida.

Details

This sound was recorded at a sampling rate of 96 kHz. The ambient temperature was 22.6 degrees Celsius. The sound file has been imported into an R object using the readWave function from the tuneR package.

Examples

if (FALSE) { # \dontrun{
data(coryphoda)
summary(coryphoda)
spectrum_plotly(coryphoda)
} # }