Oscillogram with ggplot2

Oscillogram with ggplot2

Usage

oscillogram_ggplot(
  wave,
  y_title = "Relative Amplitude",
  x_title = "",
  save = TRUE,
  file_name = "",
  show_x = TRUE
)

Arguments

wave

A Wave object

y_title

Character. A title for the Y-axis.

x_title

Character. A title for the X-axis.

save

Logical. If TRUE, the plot is saved in the working directory. Defaults to FALSE.

file_name

Character. Name for the PNG file to be saved. A suffix ("_oscillogram") is added by default.

show_x

Logical. If TRUE (default), the X-axis is plotted.

Value

An oscillogram plot. This plot is meant to be used with other plots with matching the X-axis, therefore it lacks its labels.

Examples

if (FALSE) { # \dontrun{
oscillgoram_ggplot(coryphoda)
} # }