Multiplot including Spectrogram, Mean Power Spectrum, and

Multiplot including Spectrogram, Mean Power Spectrum, and

multiplot_ggplot(
  wave,
  cutoff = -35,
  overlap = 95,
  zeropad = 200,
  scale.type = "dB",
  heights = c(7, 3),
  oscillo.label = ""
)

Arguments

wave

A `Wave` object. The sound data to be plotted.

cutoff

Numeric. Amplitude cutoff in dB below the peak. Values below the cutoff will not be visible in the spectrogram. Default is -35 dB.

overlap

Numeric. Window overlap percentage for the spectrogram. Default is 95.

zeropad

Numeric. Zero padding for FFT in the spectrogram. Default is 200.

scale.type

Character. Scale for the amplitude in the spectrogram, either "dB" or "Linear". Default is "dB".

heights

Numeric vector. The relative heights of the spectrogram/mean spectrum plot and the oscillogram. Default is `c(7,3)`.

oscillo.label

Character. A title for the Y-axis of the oscillogram. Default is an empty string.

Value

A plot with aligned spectrogram, mean spectrum, and oscillogram.

Examples

if (FALSE) { # \dontrun{
data(coryphoda)
multiplot_ggplot(coryphoda)
} # }